Create the shopping cart service
Completed the first part of the third section of the angular tutorial. Generated a cart service and defined methods to add items to the cart, remove items from the cart, and clear the cart. Modified the product details ts to allow use of the cart service on the website and added a functional buy button.